Extract computeAsyncVisualDepths to FlameGraphLayout.ts
Fixes react-refresh/only-export-components lint warning by moving the
non-component helper and its SpanData interface out of FlameGraphCell.tsx
into a sibling module. The test file now imports from the new location.
